feat: 新增对OPENCLAW和CLAUDE协议的支持
fix: 修复头文件包含错误和内存泄漏问题

refactor: 重构线程同步和IPC机制

perf: 优化定时器和事件处理性能

style: 清理代码注释和格式

test: 添加系统调用和内存管理单元测试

docs: 更新API文档和错误码定义

build: 更新CMake构建配置和依赖管理

ci: 添加持续集成测试配置

chore: 清理废弃代码和冗余实现

+++ agentos/commons/utils/ipc/include/ipc_common.h
@@ -176,7 +176,7 @@ typedef struct {
     uint64_t correlation_id;            /**< 关联 ID（请求-响应模式） */
     char source[64];                    /**< 发送者标识 */
     char target[64];                    /**< 目标标识 */
-    uint32_t payload_len;               /**< 负载长度 */
+    uint64_t payload_len;               /**< 负载长度 */
     uint32_t checksum;                  /**< 校验和 (CRC32) */
     agentos_timestamp_t timestamp;      /**< 时间戳 */
     uint8_t reserved[32];               /**< 保留字段 */
@@ -198,8 +198,8 @@ typedef struct {
     ipc_type_t type;                    /**< 通道类型 */
     const char* name;                   /**< 通道名称 */
     ipc_mode_t mode;                    /**< 读写模式 */
-    uint32_t buffer_size;               /**< 缓冲区大小 */
-    uint32_t max_message_size;          /**< 最大消息大小 */
+    size_t buffer_size;                  /**< 缓冲区大小 */
+    size_t max_message_size;             /**< 最大消息大小 */
     uint32_t timeout_ms;                /**< 默认超时 */
     uint32_t max_connections;           /**< 最大连接数（服务端） */
     bool nonblocking;                   /**< 是否非阻塞 */
